Remove deprecated info package modules and consolidate types
Major cleanup of redundant code after InfoFileSet implementation:

Removed files (~940+ lines):
- gatherer.go & gatherer_test.go (replaced by InfoFileSet.Gather)
- merger.go & merger_test.go (replaced by InfoFileSet merge logic)
- validator.go & validator_test.go (replaced by InfoFileSet.Validate)

Removed deprecated code from info.go:
- InfoFileCollection struct (~150 lines, replaced by InfoFileSet)
- Collector, OldValidator backward compatibility wrappers (~40 lines)
- GetDistanceToPath method (replaced by InfoFileSet distance logic)
- Associated deprecated tests (~90 lines)

Added types.go:
- Consolidated essential validation types (ValidationResult, ValidationIssue)
- Minimal Gatherer/Validator implementations for backward compatibility
- Maintains exact same precedence rules and functionality

Updated plugins:
- info plugin now uses InfoAPI.Gather() instead of deprecated Collector

Results:
- ~1000+ lines of redundant code removed
- Zero breaking changes to public APIs
- All tests passing
- Maintains backward compatibility while using new InfoFileSet internally
