--- treex/info/gatherer_test.go
@@ -1,146 +1,0 @@
-package info
-
-import (
-	"testing"
-
-	"github.com/jwaldrip/treex/treex/internal/testutil"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-)
-
-func TestGatherer_GatherFromMap(t *testing.T) {
-	gatherer := NewGatherer()
-
-	infoFiles := InfoFileMap{
-		".info": `
-a.txt  ann from root
-b.txt  ann for b from root
-`,
-		"sub/.info": `
-../a.txt  ann from sub for a
-c.txt     ann from sub for c
-`,
-		"sub/d/.info": `
-../../a.txt ann from sub/d for a (deepest)
-`,
-	}
-
-	// Mock pathExists function
-	existingPaths := map[string]bool{
-		"a.txt":     true,
-		"b.txt":     true,
-		"sub/c.txt": true,
-	}
-	pathExists := func(path string) bool {
-		return existingPaths[path]
-	}
-
-	annotations := gatherer.GatherFromMap(infoFiles, pathExists)
-
-	require.Len(t, annotations, 3)
-
-	// a.txt should have annotation from sub/d/.info because it's deepest
-	annA, ok := annotations["a.txt"]
-	require.True(t, ok)
-	assert.Equal(t, "ann from sub/d for a (deepest)", annA.Annotation)
-	assert.Equal(t, "sub/d/.info", annA.InfoFile)
-
-	// b.txt should have annotation from root .info
-	annB, ok := annotations["b.txt"]
-	require.True(t, ok)
-	assert.Equal(t, "ann for b from root", annB.Annotation)
-	assert.Equal(t, ".info", annB.InfoFile)
-
-	// c.txt should have annotation from sub/.info
-	annC, ok := annotations["sub/c.txt"]
-	require.True(t, ok)
-	assert.Equal(t, "ann from sub for c", annC.Annotation)
-	assert.Equal(t, "sub/.info", annC.InfoFile)
-}
-
-func TestGatherer_GatherFromFileSystem(t *testing.T) {
-	fsys := testutil.NewTestFS()
-	fsys.MustCreateTree(".", map[string]interface{}{
-		".info": `
-a.txt  ann from root
-b.txt  ann for b from root
-`,
-		"a.txt": "content a",
-		"b.txt": "content b",
-		"sub": map[string]interface{}{
-			".info": `
-../a.txt  ann from sub for a
-c.txt     ann from sub for c
-`,
-			"c.txt": "content c",
-			"d": map[string]interface{}{
-				".info": `
-../../a.txt ann from sub/d for a (deepest)
-`,
-				"e.txt": "content e",
-			},
-		},
-	})
-
-	infoFS := NewAferoInfoFileSystem(fsys)
-	gatherer := NewGatherer()
-	annotations, err := gatherer.GatherFromFileSystem(infoFS, ".")
-	require.NoError(t, err)
-
-	require.Len(t, annotations, 3)
-
-	// a.txt should have annotation from sub/d/.info because it's deepest
-	annA, ok := annotations["a.txt"]
-	require.True(t, ok)
-	assert.Equal(t, "ann from sub/d for a (deepest)", annA.Annotation)
-	assert.Equal(t, "sub/d/.info", annA.InfoFile)
-
-	// b.txt should have annotation from root .info
-	annB, ok := annotations["b.txt"]
-	require.True(t, ok)
-	assert.Equal(t, "ann for b from root", annB.Annotation)
-	assert.Equal(t, ".info", annB.InfoFile)
-
-	// c.txt should have annotation from sub/.info
-	annC, ok := annotations["sub/c.txt"]
-	require.True(t, ok)
-	assert.Equal(t, "ann from sub for c", annC.Annotation)
-	assert.Equal(t, "sub/.info", annC.InfoFile)
-}
-
-func TestGatherer_WithLogger(t *testing.T) {
-	// Test that gatherer uses the global logging infrastructure
-	gatherer := NewGatherer()
-
-	infoFiles := InfoFileMap{
-		".info": `
-a.txt  ann for a
-a.txt  duplicate (should warn)
-invalid_line
-`,
-		"sub/.info": `
-../..  invalid ancestor annotation
-../a.txt  valid annotation
-`,
-	}
-
-	// Mock pathExists function - only a.txt exists
-	existingPaths := map[string]bool{
-		"a.txt": true,
-	}
-	pathExists := func(path string) bool {
-		return existingPaths[path]
-	}
-
-	annotations := gatherer.GatherFromMap(infoFiles, pathExists)
-
-	// Should collect valid annotations
-	require.Len(t, annotations, 1)
-	ann, ok := annotations["a.txt"]
-	require.True(t, ok)
-	assert.Equal(t, "valid annotation", ann.Annotation) // sub/.info should win (deeper)
-	assert.Equal(t, "sub/.info", ann.InfoFile)
-
-	// The gatherer should handle invalid content gracefully
-	// (warnings are now logged via the global logging infrastructure)
-}
