+++ treex/info/info.go
@@ -57,11 +57,9 @@ package info
 
 import (
 	"io"
-	"path/filepath"
 	"strings"
 
 	"github.com/jwaldrip/treex/treex/logging"
-	"github.com/spf13/afero"
 )
 
 // Annotation represents a single entry in an .info file.
@@ -232,47 +230,6 @@ func (info *InfoFile) AddAnnotationForPath(path, annotation string) bool {
 	return true
 }
 
-// GetDistanceToPath calculates distance from this .info file to target path
-// Returns 0 if same directory, 1 if child, -1 if parent/outside scope
-func (info *InfoFile) GetDistanceToPath(targetPath string) int {
-	infoDir := filepath.Dir(info.Path)
-
-	// Handle relative target paths - they are relative to the info file's directory
-	var targetDir string
-	if filepath.IsAbs(targetPath) {
-		targetDir = filepath.Dir(targetPath)
-	} else {
-		// For relative paths, join with info directory first
-		fullTargetPath := filepath.Join(infoDir, targetPath)
-		targetDir = filepath.Dir(fullTargetPath)
-	}
-
-	// Make paths clean for comparison
-	infoDir = filepath.Clean(infoDir)
-	targetDir = filepath.Clean(targetDir)
-
-	if infoDir == targetDir {
-		return 0 // Same directory
-	}
-
-	// Check if target is under info directory
-	rel, err := filepath.Rel(infoDir, targetDir)
-	if err != nil {
-		return -1 // Can't calculate relationship
-	}
-
-	if strings.HasPrefix(rel, "..") {
-		return -1 // Target is outside or above info directory
-	}
-
-	// Count directory levels
-	if rel == "." {
-		return 0
-	}
-
-	return strings.Count(rel, string(filepath.Separator)) + 1
-}
-
 // IsEmpty returns true if the file has no valid annotations
 func (info *InfoFile) IsEmpty() bool {
 	return len(info.annotations) == 0
@@ -370,206 +327,4 @@ func (writer *InfoFileWriter) WriteInfoFiles(infoFiles []*InfoFile) error {
 		}
 	}
 	return nil
-}
-
-// InfoFileCollection provides operations on collections of InfoFiles.
-// All operations are in-memory and work with the parsed tree structures.
-type InfoFileCollection struct {
-	infoFiles []*InfoFile
-}
-
-// NewInfoFileCollection creates a new collection
-func NewInfoFileCollection(infoFiles []*InfoFile) *InfoFileCollection {
-	return &InfoFileCollection{infoFiles: infoFiles}
-}
-
-// Distribute redistributes annotations to their optimal .info files based on path proximity
-// Returns a new collection with annotations moved to their closest .info files
-func (collection *InfoFileCollection) Distribute() *InfoFileCollection {
-	// Create map of all annotations with their current distances
-	type annotationWithDistance struct {
-		annotation *Annotation
-		distance   int
-		sourceFile *InfoFile
-		targetFile *InfoFile
-	}
-
-	var annotationsToMove []annotationWithDistance
-
-	// For each annotation, find the best .info file
-	for _, infoFile := range collection.infoFiles {
-		for _, annotation := range infoFile.GetAllAnnotations() {
-			bestFile := infoFile
-			bestDistance := infoFile.GetDistanceToPath(annotation.Path)
-
-			// Find a better file if one exists
-			for _, otherFile := range collection.infoFiles {
-				distance := otherFile.GetDistanceToPath(annotation.Path)
-				if distance >= 0 && (bestDistance < 0 || distance < bestDistance) {
-					bestFile = otherFile
-					bestDistance = distance
-				}
-			}
-
-			// If we found a better file, mark for moving
-			if bestFile != infoFile {
-				annotationsToMove = append(annotationsToMove, annotationWithDistance{
-					annotation: &annotation,
-					distance:   bestDistance,
-					sourceFile: infoFile,
-					targetFile: bestFile,
-				})
-			}
-		}
-	}
-
-	// Create new collection with moved annotations
-	newFiles := make([]*InfoFile, len(collection.infoFiles))
-	for i, infoFile := range collection.infoFiles {
-		// Start with a copy of the original file structure
-		newFiles[i] = &InfoFile{
-			Path:        infoFile.Path,
-			Lines:       make([]Line, len(infoFile.Lines)),
-			annotations: make(map[string]*Annotation),
-		}
-		copy(newFiles[i].Lines, infoFile.Lines)
-
-		// Copy annotations that aren't being moved
-		for path, ann := range infoFile.annotations {
-			shouldMove := false
-			for _, move := range annotationsToMove {
-				if move.annotation.Path == path && move.sourceFile == infoFile {
-					shouldMove = true
-					break
-				}
-			}
-			if !shouldMove {
-				newFiles[i].annotations[path] = ann
-			} else {
-				// Remove from lines by marking as removed
-				newFiles[i].RemoveAnnotationForPath(path)
-			}
-		}
-	}
-
-	// Add moved annotations to their target files
-	for _, move := range annotationsToMove {
-		for _, newFile := range newFiles {
-			if newFile.Path == move.targetFile.Path {
-				newFile.AddAnnotationForPath(move.annotation.Path, move.annotation.Annotation)
-				break
-			}
-		}
-	}
-
-	return NewInfoFileCollection(newFiles)
-}
-
-// Gather merges all annotations into a single .info file at the root
-// Returns a new collection with a single .info file containing all annotations
-func (collection *InfoFileCollection) Gather(rootPath string) *InfoFileCollection {
-	rootInfoPath := filepath.Join(rootPath, ".info")
-
-	// Create a new root info file
-	rootInfo := &InfoFile{
-		Path:        rootInfoPath,
-		Lines:       []Line{},
-		annotations: make(map[string]*Annotation),
-	}
-
-	// Collect all annotations from all files
-	for _, infoFile := range collection.infoFiles {
-		for _, annotation := range infoFile.GetAllAnnotations() {
-			// Avoid duplicates (first one wins as per spec)
-			if !rootInfo.HasAnnotationForPath(annotation.Path) {
-				rootInfo.AddAnnotationForPath(annotation.Path, annotation.Annotation)
-			}
-		}
-	}
-
-	// Create empty versions of other info files for removal
-	var newFiles []*InfoFile
-	newFiles = append(newFiles, rootInfo)
-
-	for _, infoFile := range collection.infoFiles {
-		if infoFile.Path != rootInfoPath {
-			emptyFile := &InfoFile{
-				Path:        infoFile.Path,
-				Lines:       []Line{},
-				annotations: make(map[string]*Annotation),
-			}
-			newFiles = append(newFiles, emptyFile)
-		}
-	}
-
-	return NewInfoFileCollection(newFiles)
-}
-
-// GetFilesToRemove returns files that are empty and should be removed
-func (collection *InfoFileCollection) GetFilesToRemove() []string {
-	var toRemove []string
-	for _, infoFile := range collection.infoFiles {
-		if infoFile.IsEmpty() {
-			toRemove = append(toRemove, infoFile.Path)
-		}
-	}
-	return toRemove
-}
-
-// GetFiles returns all InfoFiles in the collection
-func (collection *InfoFileCollection) GetFiles() []*InfoFile {
-	return collection.infoFiles
-}
-
-// Backward compatibility functions - delegate to new modules
-
-// Collector manages the collection and merging of annotations.
-// Deprecated: Use InfoAPI or Gatherer directly for new code.
-type Collector struct {
-	gatherer *Gatherer
-}
-
-// NewCollector creates a new annotation collector.
-func NewCollector() *Collector {
-	return &Collector{
-		gatherer: NewGatherer(),
-	}
-}
-
-// CollectAnnotations walks the filesystem from root, finds all .info files,
-// parses them, and returns a map of path to the winning annotation.
-func (c *Collector) CollectAnnotations(fsys afero.Fs, root string) (map[string]Annotation, error) {
-	infoFS := NewAferoInfoFileSystem(fsys)
-	return c.gatherer.GatherFromFileSystem(infoFS, root)
-}
-
-// OldValidator provides validation functionality for .info files
-// Deprecated: Use InfoAPI.Validate() or Validator directly for new code.
-type OldValidator struct {
-	validator *Validator
-}
-
-// NewValidator creates a new info file validator
-// Deprecated: Use NewInfoValidator() from validator.go or InfoAPI for new code.
-func NewValidator(fs afero.Fs) *OldValidator {
-	return &OldValidator{
-		validator: NewInfoValidator(),
-	}
-}
-
-// NewValidatorWithLogger creates a new info file validator with a custom logger
-// Deprecated: Use NewInfoValidator() from validator.go or InfoAPI for new code.
-// The logger parameter is ignored - logging now uses the global logging infrastructure.
-func NewValidatorWithLogger(fs afero.Fs, logger interface{}) *OldValidator {
-	return &OldValidator{
-		validator: NewInfoValidator(),
-	}
-}
-
-// ValidateDirectory validates all .info files in a directory tree
-func (v *OldValidator) ValidateDirectory(rootPath string) (*ValidationResult, error) {
-	// Note: The original OldValidator had an fs field, but for backward compatibility
-	// we'll need the filesystem to be passed in or use the OS filesystem
-	// For now, this is a placeholder that would need the filesystem passed through
-	panic("OldValidator.ValidateDirectory needs filesystem - use InfoAPI instead")
 }
