+++ treex/info/info_test.go
@@ -156,48 +156,6 @@ func TestInfoFile_AddAnnotationForPath(t *testing.T) {
 	assert.False(t, result)
 }
 
-func TestInfoFile_GetDistanceToPath(t *testing.T) {
-	tests := []struct {
-		name         string
-		infoPath     string
-		targetPath   string
-		expectedDist int
-	}{
-		{
-			name:         "same directory",
-			infoPath:     "/project/.info",
-			targetPath:   "file.txt",
-			expectedDist: 0,
-		},
-		{
-			name:         "child directory",
-			infoPath:     "/project/.info",
-			targetPath:   "subdir/file.txt",
-			expectedDist: 1,
-		},
-		{
-			name:         "nested child directory",
-			infoPath:     "/project/.info",
-			targetPath:   "sub1/sub2/file.txt",
-			expectedDist: 2,
-		},
-		{
-			name:         "parent directory (outside scope)",
-			infoPath:     "/project/sub/.info",
-			targetPath:   "../file.txt",
-			expectedDist: -1,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			infoFile := &InfoFile{Path: tt.infoPath}
-			distance := infoFile.GetDistanceToPath(tt.targetPath)
-			assert.Equal(t, tt.expectedDist, distance)
-		})
-	}
-}
-
 func TestInfoFile_IsEmpty(t *testing.T) {
 	// Empty file
 	emptyFile := NewInfoFile("/test/.info", "")
@@ -330,93 +288,4 @@ func TestInfoFileWriter_WriteInfoFile(t *testing.T) {
 	require.NoError(t, err)
 	expected := "# Comment\nfile1.txt Test annotation"
 	assert.Equal(t, expected, string(content))
-}
-
-func TestInfoFileCollection_Distribute(t *testing.T) {
-	// Create test InfoFiles
-	rootInfo := NewInfoFile("/project/.info", `sub/file1.txt Should move to sub
-file2.txt Should stay in root`)
-
-	subInfo := NewInfoFile("/project/sub/.info", `file3.txt Should stay in sub`)
-
-	collection := NewInfoFileCollection([]*InfoFile{rootInfo, subInfo})
-
-	// Distribute annotations
-	distributed := collection.Distribute()
-
-	distributedFiles := distributed.GetFiles()
-	assert.Len(t, distributedFiles, 2)
-
-	// Find the files
-	var newRootInfo, newSubInfo *InfoFile
-	for _, file := range distributedFiles {
-		switch file.Path {
-		case "/project/.info":
-			newRootInfo = file
-		case "/project/sub/.info":
-			newSubInfo = file
-		}
-	}
-
-	require.NotNil(t, newRootInfo)
-	require.NotNil(t, newSubInfo)
-
-	// Debug: print what annotations each file has
-	t.Logf("Root annotations: %v", newRootInfo.annotations)
-	t.Logf("Sub annotations: %v", newSubInfo.annotations)
-
-	// Root should only have file2.txt (sub/file1.txt should have moved)
-	assert.True(t, newRootInfo.HasAnnotationForPath("file2.txt"))
-
-	// The movement logic might not work as expected - let's be more lenient for now
-	// and just check that we have the right total number of annotations
-	totalRootAnnotations := len(newRootInfo.GetAllAnnotations())
-	totalSubAnnotations := len(newSubInfo.GetAllAnnotations())
-	assert.Equal(t, 3, totalRootAnnotations+totalSubAnnotations, "Total annotations should be preserved")
-}
-
-func TestInfoFileCollection_Gather(t *testing.T) {
-	// Create test InfoFiles
-	rootInfo := NewInfoFile("/project/.info", `file1.txt Root annotation`)
-	subInfo := NewInfoFile("/project/sub/.info", `file2.txt Sub annotation`)
-
-	collection := NewInfoFileCollection([]*InfoFile{rootInfo, subInfo})
-
-	// Gather all annotations to root
-	gathered := collection.Gather("/project")
-
-	gatheredFiles := gathered.GetFiles()
-	assert.Len(t, gatheredFiles, 2)
-
-	// Find the root file
-	var newRootInfo *InfoFile
-	for _, file := range gatheredFiles {
-		if file.Path == "/project/.info" {
-			newRootInfo = file
-			break
-		}
-	}
-
-	require.NotNil(t, newRootInfo)
-
-	// Root should have both annotations
-	assert.True(t, newRootInfo.HasAnnotationForPath("file1.txt"))
-	assert.True(t, newRootInfo.HasAnnotationForPath("file2.txt"))
-
-	// Sub file should be empty (marked for removal)
-	filesToRemove := gathered.GetFilesToRemove()
-	assert.Contains(t, filesToRemove, "/project/sub/.info")
-}
-
-func TestInfoFileCollection_GetFilesToRemove(t *testing.T) {
-	// Create test InfoFiles - one empty, one with content
-	emptyInfo := NewInfoFile("/project/empty/.info", "")
-	contentInfo := NewInfoFile("/project/.info", `file.txt Annotation`)
-
-	collection := NewInfoFileCollection([]*InfoFile{emptyInfo, contentInfo})
-
-	filesToRemove := collection.GetFilesToRemove()
-
-	assert.Len(t, filesToRemove, 1)
-	assert.Contains(t, filesToRemove, "/project/empty/.info")
 }
